split use-user hook test into unit and integration
